Test cleanups to eliminate duplicate filenames
